Introduction

Alphabet

Grammar

Algebra

Logic

More about Connectives

Boolean logic

Introduction

At the end of this chapter, you are going to understand how Atoms and Truth values are organized in alphabets, and how grammars are defined.

You will also get an in-depth view of how boolean truth values and boolean Connective's are defined from both a syntax and a syntactical standpoint of view.

The end of this chapter is dedicated to modal logic, which is one of the most hot topics covered by SoleLogics.

Recalling the type hierarchy presented in man-core, it is here enriched with the following new types and structures.

Truth

BooleanTruth (new)

Top

Bot

Connective

NamedConnective (new)

NEGATION

CONJUNCTION

DISJUNCTION

IMPLICATION

DIAMOND

BOX

AbstractAlphabet{V} (new)

ExplicitAlphabet{V}

AlphabetOfAny{V}

AbstractGrammar{V<:AbstractAlphabet,O<:Operator} (new)

CompleteFlatGrammar{V<:AbstractAlphabet,O<:Operator}

AbstractAlgebra{T<:Truth} (new)

BooleanAlgebra

AbstractLogic{G<:AbstractGrammar,V<:AbstractAlgebra} (new)

[BaseLogic{G<:AbstractGrammar,A<:AbstractAlgebra}]

Alphabet

SoleLogics.AbstractAlphabet — Type

abstract type AbstractAlphabet{V} end

Abstract type for representing an alphabet of atoms with values of type V. An alphabet (or propositional alphabet) is a set of atoms (assumed to be countable).

Examples

julia> Atom(1) in ExplicitAlphabet(Atom.(1:10))
true

julia> Atom(1) in ExplicitAlphabet(1:10)
true

julia> Atom(1) in AlphabetOfAny{String}()
false

julia> Atom("mystring") in AlphabetOfAny{String}()
true

julia> "mystring" in AlphabetOfAny{String}()
┌ Warning: Please, use Base.in(Atom(mystring), alphabet::AlphabetOfAny{String}) instead of Base.in(mystring, alphabet::AlphabetOfAny{String})
└ @ SoleLogics ...
true

Implementation

When implementing a new alphabet type MyAlphabet, you should provide a method for establishing whether an atom belongs to it or not; while, in general, this method should be:

function Base.in(p::Atom, a::MyAlphabet)::Bool

in the case of finite alphabets, it suffices to define a method:

function atoms(a::AbstractAlphabet)::AbstractVector{atomstype(a)}

By default, an alphabet is considered finite:

Base.isfinite(::Type{<:AbstractAlphabet}) = true
Base.isfinite(a::AbstractAlphabet) = Base.isfinite(typeof(a))
Base.in(p::Atom, a::AbstractAlphabet) = Base.isfinite(a) ? Base.in(p, atoms(a)) : error(...)

See also AbstractGrammar, AlphabetOfAny, Atom, ExplicitAlphabet.

Base.isfinite — Method

Base.isfinite(a::AbstractAlphabet)

Return true if the alphabet is finite, false otherwise.

See AbstractAlphabet.

SoleLogics.atoms — Method

atoms(a::AbstractAlphabet)::AbstractVector{atomstype(a)}

List the atoms of a finite alphabet.

See also AbstractAlphabet.

Base.in — Method

Base.in(p::Atom, a::AbstractAlphabet)::Bool

Return whether an atom belongs to an alphabet.

See also AbstractAlphabet, Atom.

Base.length — Method

Base.length(a::AbstractAlphabet)::Bool

Return the alphabet length, if it is finite.

See also AbstractAlphabet, SyntaxBranch.

Base.iterate — Method

Base.iterate(a::AbstractAlphabet)
Base.iterate(a::AbstractAlphabet, state)

Return an iterator to the next element in an alhabet.

See also AbstractAlphabet, SyntaxBranch.

SoleLogics.ExplicitAlphabet — Type

struct ExplicitAlphabet{V} <: AbstractAlphabet{V}
    atoms::Vector{Atom{V}}
end

An alphabet wrapping atoms in a (finite) Vector.

See also AbstractAlphabet, atoms.

SoleLogics.AlphabetOfAny — Type

struct AlphabetOfAny{V} <: AbstractAlphabet{V} end

An implicit, infinite alphabet that includes all atoms with values of a subtype of V.

See also AbstractAlphabet.

Grammar

SoleLogics.AbstractGrammar — Type

abstract type AbstractGrammar{V<:AbstractAlphabet,O<:Operator} end

Abstract type for representing a context-free grammar based on a single alphabet of type V, and a set of operators that consists of all the (singleton) child types of O. V context-free grammar is a simple structure for defining formulas inductively.

See also alphabet, AbstractAlphabet, Operator.

SoleLogics.alphabet — Method

alphabet(g::AbstractGrammar{V} where {V})::V

Return the propositional alphabet of a grammar.

See also AbstractAlphabet, AbstractGrammar.

Base.in — Method

Base.in(φ::SyntaxTree, g::AbstractGrammar)::Bool

Return whether a SyntaxTree, belongs to a grammar.

See also AbstractGrammar, SyntaxTree.

SoleLogics.formulas — Method

formulas(
    g::AbstractGrammar;
    maxdepth::Integer,
    nformulas::Union{Nothing,Integer} = nothing,
    args...
)::Vector{<:SyntaxBranch}

Enumerate the formulas produced by a given grammar with a finite and iterable alphabet.

Implementation

Additional args can be used to model the function's behavior. At least these two arguments should be covered:

a nformulas argument can be used to limit the size of the returned Vector;

a maxdepth argument can be used to limit the syntactic component, represented as a syntax tree,

to a given maximum depth;

See also AbstractGrammar, SyntaxBranch.

SoleLogics.CompleteFlatGrammar — Type

struct CompleteFlatGrammar{V<:AbstractAlphabet,O<:Operator} <: AbstractGrammar{V,O}
    alphabet::V
    operators::Vector{<:O}
end

V grammar of all well-formed formulas obtained by the arity-complying composition of atoms of an alphabet of type V, and all operators in operators. With n operators, this grammar has exactly n+1 production rules. For example, with operators = [∧,∨], the grammar (in Backus-Naur form) is:

φ ::= p | φ ∧ φ | φ ∨ φ

with p ∈ alphabet. Note: it is flat in the sense that all rules substitute the same (unique and starting) non-terminal symbol φ.

See also AbstractGrammar, Operator, alphabet, formulas, connectives, operators, leaves.

SoleLogics.connectives — Method

connectives(g::AbstractGrammar)

List all connectives appearing in a grammar.

See also Connective, nconnectives.

SoleLogics.leaves — Method

leaves(g::AbstractGrammar)

List all leaves appearing in a grammar.

See also SyntaxLeaf, nleaves.

SoleLogics.formulas — Method

formulas(
    g::CompleteFlatGrammar{V,O} where {V,O};
    maxdepth::Integer,
    nformulas::Union{Nothing,Integer} = nothing
)::Vector{SyntaxBranch}

Generate all formulas whose SyntaxBranchs that are not taller than a given maxdepth.

See also AbstractGrammar, SyntaxBranch.

Algebra

SoleLogics.AbstractAlgebra — Type

abstract type AbstractAlgebra{T<:Truth} end

Abstract type for representing algebras. Algebras are used for grounding the truth of atoms and the semantics of operators. They typically encode a lattice structure where two elements(or nodes) ⊤ and ⊥ are referred to as TOP (or maximum) and bot (or minimum). Each node in the lattice represents a truth value that an atom or a formula can have on an interpretation, and the semantics of operators is given in terms of operations between truth values.

Implementation

When implementing a new algebra type, the methods domain, TOP, and bot should be implemented.

See also bot, BooleanAlgebra, Operator, TOP, collatetruth, domain, iscrisp, truthtype.

SoleLogics.truthtype — Method

truthtype(::Type{<:AbstractAlgebra{T}}) where {T<:Truth} = T
truthtype(a::AbstractAlgebra) = truthtype(typeof(a))

The Julia type for representing truth values of the algebra.

See also AbstractAlgebra.

SoleLogics.domain — Method

domain(a::AbstractAlgebra)

Return an iterator to the values in the domain of a given algebra.

See also AbstractAlgebra.

SoleLogics.top — Method

top(a::AbstractAlgebra)

Return the top of a given algebra.

See also bot, AbstractAlgebra.

SoleLogics.bot — Method

bot(a::AbstractAlgebra)

Return the bottom of a given algebra.

See also top, AbstractAlgebra.

SoleLogics.iscrisp — Method

iscrisp(a::AbstractAlgebra) = iscrisp(typeof(a))

An algebra is crisp (or boolean) when its domain only has two values, namely, the top and the bottom. The antonym of crisp is fuzzy.

See also AbstractAlgebra.

Logic

SoleLogics.AbstractLogic — Type

abstract type AbstractLogic{G<:AbstractGrammar,V<:AbstractAlgebra} end

Abstract type of a logic, which comprehends a context-free grammar (syntax) and an algebra (semantics).

Implementation

When implementing a new logic type, the methods grammar and algebra should be implemented.

See also AbstractAlgebra, AbstractGrammar.

SoleLogics.grammar — Method

grammar(l::AbstractLogic{G})::G where {G<:AbstractGrammar}

Return the grammar of a given logic.

See also AbstractGrammar, AbstractLogic, algebra, alphabet, formulas, grammar, operators, truthtype.

SoleLogics.algebra — Method

algebra(l::AbstractLogic{G,V})::V where {G,V}

Return the algebra of a given logic.

See also AbstractAlgebra, AbstractLogic.

More about Connectives

SoleLogics.NamedConnective — Type

struct NamedConnective{Symbol} <: Connective end

A singleton type for representing connectives defined by a name or a symbol.

Examples

The AND connective (i.e., the logical conjunction) is defined as the subtype:

const CONJUNCTION = NamedConnective{:∧}()
const ∧ = CONJUNCTION
arity(::typeof(∧)) = 2

See also NEGATION, CONJUNCTION, DISJUNCTION, IMPLICATION, Connective.

SoleLogics.collatetruth — Method

collatetruth(c::Connective, ts::NTuple{N,T where T<:Truth})::Truth where {N}

Return the truth value for a composed formula c(t1, ..., tN), given the N with t1, ..., tN being Truth values.

See also simplify, Connective, Truth.

SoleLogics.simplify — Method

simplify(c::Connective, ts::NTuple{N,F where F<:Formula})::Truth where {N}

Return a formula with the same semantics of a composed formula c(φ1, ..., φN), given the N immediate sub-formulas.

See also collatetruth, Connective, Formula.

Boolean logic

SoleLogics.NEGATION — Constant

const NEGATION = NamedConnective{:¬}()
const ¬ = NEGATION
arity(::typeof(¬)) = 1

Logical negation (also referred to as complement). It can be typed by \neg<tab>.

See also NamedConnective, Connective.

SoleLogics.CONJUNCTION — Constant

const CONJUNCTION = NamedConnective{:∧}()
const ∧ = CONJUNCTION
arity(::typeof(∧)) = 2

Logical conjunction. It can be typed by \wedge<tab>.

See also NamedConnective, Connective.

SoleLogics.DISJUNCTION — Constant

const DISJUNCTION = NamedConnective{:∨}()
const ∨ = DISJUNCTION
arity(::typeof(∨)) = 2

Logical disjunction. It can be typed by \vee<tab>.

See also NamedConnective, Connective.

SoleLogics.IMPLICATION — Constant

const IMPLICATION = NamedConnective{:→}()
const → = IMPLICATION
arity(::typeof(→)) = 2

Logical implication. It can be typed by \to<tab>.

See also NamedConnective, Connective.

Boolean logic Connectives are regrouped in a single collection.

SoleLogics.BASE_CONNECTIVES — Constant

const BASE_CONNECTIVES = [¬, ∧, ∨, →]

Basic logical operators.

See also NEGATION, CONJUNCTION, DISJUNCTION, IMPLICATION, Connective.

SoleLogics.BooleanTruth — Type

abstract type BooleanTruth <: Truth end

Supertype of Top and Bot, the two truth values of BooleanAlgebra

See also Bot, Top, BooleanAlgebra.

SoleLogics.Top — Type

struct Top <: Truth end
const TOP = Top()
const ⊤ = TOP

Canonical truth operator representing the value true. It can be typed by \top<tab>.

See also BOT, Truth.

SoleLogics.Bot — Type

struct Bot <: Truth end
const BOT = Bot()
const ⊥ = BOT

Canonical truth operator representing the value false. It can be typed by \bot<tab>.

See also TOP, Truth.

SoleLogics.BooleanAlgebra — Type

struct BooleanAlgebra <: AbstractAlgebra{Bool} end

A Boolean algebra, defined on the values Top (representing true) and Bot (for bottom, representing false). For this algebra, the basic operators negation, conjunction and disjunction (stylized as ¬, ∧, ∨) can be defined as the complement, minimum and maximum, of the integer cast of true and false, respectively.

See also Truth.

SoleLogics.BaseLogic — Type

struct BaseLogic{G<:AbstractGrammar,A<:AbstractAlgebra} <: AbstractLogic{G,A}
    grammar::G
    algebra::A
end

A basic logic based on a grammar and an algebra, where both the grammar and the algebra are instantiated.

See also grammar, algebra, AbstractGrammar, AbstractAlgebra, AbstractLogic.