Alphabets

SoleLogics.AbstractAlphabet — Type

abstract type AbstractAlphabet{A} end

Abstract type for representing an alphabet of propositions with atoms of type A. An alphabet (or propositional alphabet) is a set of propositions, and it is assumed to be countable.

See also ExplicitAlphabet, AlphabetOfAny, propositionstype, atomtype, Proposition, AbstractGrammar.

Examples

julia> Proposition(1) in ExplicitAlphabet(Proposition.(1:10))
true

julia> Proposition(1) in ExplicitAlphabet(1:10)
true

julia> Proposition(1) in AlphabetOfAny{String}()
false

julia> Proposition("mystring") in AlphabetOfAny{String}()
true

julia> "mystring" in AlphabetOfAny{String}()
┌ Warning: Please, use Base.in(Proposition(mystring), alphabet::AlphabetOfAny{String}) instead of Base.in(mystring, alphabet::AlphabetOfAny{String})
└ @ SoleLogics ...
true

Implementation

When implementing a new alphabet type MyAlphabet, you should provide a method for establishing whether a proposition belongs to it or not; while, in general, this method should be:

function Base.in(p::Proposition, a::MyAlphabet)::Bool

in the case of finite alphabets, it suffices to define a method:

function propositions(a::AbstractAlphabet)::AbstractVector{propositionstype(a)}

By default, an alphabet is considered finite:

Base.isfinite(::Type{<:AbstractAlphabet}) = true
Base.isfinite(a::AbstractAlphabet) = Base.isfinite(typeof(a))
Base.in(p::Proposition, a::AbstractAlphabet) = Base.isfinite(a) ? Base.in(p, propositions(a)) : error(...)

Base.in — Function

Base.in(tok::AbstractSyntaxToken, f::AbstractFormula)::Bool

Return whether a syntax token appears in a formula.

See also AbstractSyntaxToken.

Base.in(tok::AbstractSyntaxToken, tree::SyntaxTree)::Bool

Return whether a token appears in a syntax tree or not.

See also tokens, SyntaxTree.

Base.in(t::SyntaxTree, g::AbstractGrammar)::Bool

Return whether a formula, encoded as a SyntaxTree, belongs to a grammar.

See also AbstractGrammar, SyntaxTree.

Missing docstring.

Missing docstring for Base.isiterable. Check Documenter's build log for details.

Missing docstring.

Missing docstring for Base.isfinite. Check Documenter's build log for details.

SoleLogics.ExplicitAlphabet — Type

struct ExplicitAlphabet{A} <: AbstractAlphabet{A}
    propositions::Vector{Proposition{A}}
end

An alphabet wrapping propositions in a (finite) Vector.

See also propositions, AbstractAlphabet.

SoleLogics.AlphabetOfAny — Type

struct AlphabetOfAny{A} <: AbstractAlphabet{A} end

An implicit, infinite alphabet that includes all propositions with atoms of a subtype of A.

See also AbstractAlphabet.