Formulas

SoleLogics.AbstractFormula — Type

abstract type AbstractFormula end

A logical formula encoding a statement which truth can be evaluated on interpretations (or models) of the logic.

Its syntactic component is canonically encoded via a syntax tree (see SyntaxTree), and it can be anchored to a logic (see Formula).

Implementation

When implementing a new formula type MyCustomFormulaType, please provide the method for extracting its syntax tree representation:

function tree(f::MyCustomFormulaType)::SyntaxTree
    ...
end

As well as the one used for composing formulas:

function (op::AbstractOperator)(
    children::NTuple{N,Union{AbstractSyntaxToken,MyCustomFormulaType}},
)::AbstractFormula where {N}
    # Composed formula
end

See also Formula, SyntaxTree, AbstractSyntaxStructure, AbstractLogic.

SoleLogics.joinformulas — Method

joinformulas(
    op::AbstractOperator,
    ::NTuple{N,F}
)::F where {N,F<:AbstractFormula}

Return a new formula of type F by composing N formulas of the same type via an operator op. This function provides a way for composing formulas, but it allows to use operators for a more flexible composition; see the examples (and more in the extended help).

Examples

julia> f = parseformula("◊(p→q)");

julia> p = Proposition("p");

julia> syntaxstring(∧(f, p))
"(◊(p → q)) ∧ p"

julia> syntaxstring(f ∧ ¬p) # Leverage infix notation ;)
"(◊(p → q)) ∧ (¬(p))"

julia> syntaxstring(∧(f, p, ¬p))
"(◊(p → q)) ∧ (p ∧ (¬(p)))"

Implementation

Upon joinformulas lies a more flexible way of using operators for composing formulas and syntax tokens (e.g., propositions), given methods like the following:

function (op::AbstractOperator)(
    children::NTuple{N,Union{AbstractSyntaxToken,AbstractFormula}},
) where {N}
    ...
end

These allow composing formulas as in ∧(f, ¬p), and in order to access this composition with any newly defined subtype of AbstractFormula, a new method for joinformulas should be defined, together with promotion from/to other AbstractFormulas should be taken care of (see here and here).

Similarly, for allowing a (possibly newly defined) operator to be applied on a number of syntax tokens/formulas that differs from its arity, for any newly defined operator op, new methods similar to the two above should be defined. For example, although ∧ and ∨ are binary, (i.e., have arity equal to 2), compositions such as ∧(f, f2, f3, ...) and ∨(f, f2, f3, ...) can be done thanks to the following two methods that were defined in SoleLogics:

function ∧(
    c1::Union{AbstractSyntaxToken,AbstractFormula},
    c2::Union{AbstractSyntaxToken,AbstractFormula},
    c3::Union{AbstractSyntaxToken,AbstractFormula},
    cs::Union{AbstractSyntaxToken,AbstractFormula}...
)
    return ∧(c1, ∧(c2, c3, cs...))
end
function ∨(
    c1::Union{AbstractSyntaxToken,AbstractFormula},
    c2::Union{AbstractSyntaxToken,AbstractFormula},
    c3::Union{AbstractSyntaxToken,AbstractFormula},
    cs::Union{AbstractSyntaxToken,AbstractFormula}...
)
    return ∨(c1, ∨(c2, c3, cs...))
end

See also AbstractFormula, AbstractOperator.

Missing docstring.

Missing docstring for tokens(f::SoleLogics.AbstractFormula)::AbstractVector{<:SoleLogics.AbstractSyntaxToken}. Check Documenter's build log for details.

SoleLogics.operators — Method

tokens(f::AbstractFormula)::AbstractVector{<:AbstractSyntaxToken}
operators(f::AbstractFormula)::AbstractVector{<:AbstractOperator}
propositions(f::AbstractFormula)::AbstractVector{<:Proposition}
ntokens(f::AbstractFormula)::Integer
noperators(f::AbstractFormula)::Integer
npropositions(f::AbstractFormula)::Integer

Return the list or the number of (unique) syntax tokens/operators/propositions appearing in a formula.

See also AbstractSyntaxStructure.

SoleLogics.propositions — Method

tokens(f::AbstractFormula)::AbstractVector{<:AbstractSyntaxToken}
operators(f::AbstractFormula)::AbstractVector{<:AbstractOperator}
propositions(f::AbstractFormula)::AbstractVector{<:Proposition}
ntokens(f::AbstractFormula)::Integer
noperators(f::AbstractFormula)::Integer
npropositions(f::AbstractFormula)::Integer

Return the list or the number of (unique) syntax tokens/operators/propositions appearing in a formula.

See also AbstractSyntaxStructure.

Missing docstring.

Missing docstring for ntokens(f::SoleLogics.AbstractFormula). Check Documenter's build log for details.

Missing docstring.

Missing docstring for npropositions(f::SoleLogics.AbstractFormula). Check Documenter's build log for details.

Missing docstring.

Missing docstring for height. Check Documenter's build log for details.