Grammar

SoleLogics.AbstractGrammar — Type

abstract type AbstractGrammar{A<:AbstractAlphabet,O<:AbstractOperator} end

Abstract type for representing a context-free grammar based on a single alphabet of type A, and a set of operators that consists of all the (singleton) child types of O. A context-free grammar is a simple structure for defining formulas inductively.

See also alphabet, propositionstype, tokenstype, operatorstype, alphabettype, AbstractAlphabet, AbstractOperator.

SoleLogics.alphabet — Method

alphabet(g::AbstractGrammar{A} where {A})::A

Return the propositional alphabet of a grammar.

See also AbstractAlphabet, AbstractGrammar.

Missing docstring.

Missing docstring for Base.in(tok::SoleLogics.AbstractSyntaxToken, g::SoleLogics.AbstractGrammar). Check Documenter's build log for details.

SoleLogics.propositions — Function

tokens(f::AbstractFormula)::AbstractVector{<:AbstractSyntaxToken}
operators(f::AbstractFormula)::AbstractVector{<:AbstractOperator}
propositions(f::AbstractFormula)::AbstractVector{<:Proposition}
ntokens(f::AbstractFormula)::Integer
noperators(f::AbstractFormula)::Integer
npropositions(f::AbstractFormula)::Integer

Return the list or the number of (unique) syntax tokens/operators/propositions appearing in a formula.

See also AbstractSyntaxStructure.

propositions(t::SyntaxTree)::AbstractVector{Proposition}

List all propositions appearing in a syntax tree.

See also npropositions, operators, tokens, Proposition.

propositions(a::AbstractAlphabet)::AbstractVector{propositionstype(a)}

List the propositions of a finite alphabet.

See also AbstractAlphabet, Base.isfinite.