Operators

SoleLogics.AbstractOperator — Type

abstract type AbstractOperator <: AbstractSyntaxToken end

An operator is a logical constant which establishes a relation between propositions (i.e., facts). For example, the boolean operators AND, OR and IMPLIES (stylized as ∧, ∨ and →) are used to connect propositions and/or formulas to express derived concepts.

Since operators display very different algorithmic behaviors, all structs that are subtypes of AbstractOperator must be parametric singleton types, which can be dispatched upon.

Implementation

When implementing a new type for a commutative operator O with arity higher than 1, please provide a method _iscommutative, such as:

_iscommutative(::Type{O}) = true

This can help model checking operations.

See also AbstractSyntaxToken, NamedOperator, check.

SoleLogics.iscommutative — Method

iscommutative(::Type{AbstractOperator}) = false
iscommutative(o::AbstractOperator) = iscommutative(typeof(o))

Return whether it is known that an AbstractOperator is commutative.

Examples

julia> iscommutative(∧)
true
julia> iscommutative(→)
false

See also isunary, isnullary.

Extended help

Since nullary and unary operators are always commutative, this function is actually implemented as:

iscommutative(O::Type{<:AbstractOperator}) = isnullary(O) || isunary(O) || _iscommutative(O)
iscommutative(o::AbstractOperator) = iscommutative(typeof(o))
_iscommutative(::Type{<:AbstractOperator}) = false

SoleLogics.operators — Function

tokens(f::AbstractFormula)::AbstractVector{<:AbstractSyntaxToken}
operators(f::AbstractFormula)::AbstractVector{<:AbstractOperator}
propositions(f::AbstractFormula)::AbstractVector{<:Proposition}
ntokens(f::AbstractFormula)::Integer
noperators(f::AbstractFormula)::Integer
npropositions(f::AbstractFormula)::Integer

Return the list or the number of (unique) syntax tokens/operators/propositions appearing in a formula.

See also AbstractSyntaxStructure.

operators(t::SyntaxTree)::AbstractVector{AbstractOperator}

List all operators appearing in a syntax tree.

See also noperators, propositions, tokens, AbstractOperator.

Missing docstring.

Missing docstring for noperators. Check Documenter's build log for details.