Syntax

SoleLogics.AbstractSyntaxToken — Type

abstract type AbstractSyntaxToken end

A token in a syntactic structure, most commonly, a syntax tree. A syntax tree is a tree-like structure representing a logical formula, where each node holds a token, and has as many children as the arity of the token.

See also SyntaxTree, AbstractSyntaxStructure, arity, syntaxstring.

SoleLogics.arity — Method

arity(::Type{<:AbstractSyntaxToken})::Integer
arity(tok::AbstractSyntaxToken)::Integer = arity(typeof(tok))

Return the arity of a syntax token. The arity of a syntax token is an integer representing the number of allowed children in a SyntaxTree. Tokens with arity equal to 0, 1 or 2 are called nullary, unary and binary, respectively.

See also AbstractSyntaxToken.

SoleLogics.syntaxstring — Method

syntaxstring(φ::AbstractFormula; kwargs...)::String
syntaxstring(tok::AbstractSyntaxToken; kwargs...)::String

Produces the string representation of a formula or syntax token by performing a tree traversal of the syntax tree representation of the formula. Note that this representation may introduce redundant parentheses. kwargs can be used to specify how to display syntax tokens/trees under some specific conditions.

The following kwargs are currently supported:

function_notation = false::Bool: when set to true, it forces the use of

function notation for binary operators. See here.

remove_redundant_parentheses = true::Bool: when set to false, it prints a syntaxstring

where each syntactical element is wrapped in parentheses.

Examples

julia> syntaxstring((parsebaseformula("◊((p∧s)→q)")))
"◊(p ∧ s → q)"

julia> syntaxstring((parsebaseformula("◊((p∧s)→q)")); function_notation = true)
"→(◊(∧(p, s)), q)"

See also parsebaseformula, parsetree, SyntaxTree, AbstractSyntaxToken.

Implementation

In the case of a syntax tree, syntaxstring is a recursive function that calls itself on the syntax children of each node. For a correct functioning, the syntaxstring must be defined (including kwargs...) for every newly defined AbstractSyntaxToken (e.g., operators and Propositions), in a way that it produces a unique string representation, since Base.hash and Base.isequal, at least for SyntaxTrees, rely on it.

In particular, for the case of Propositions, the function calls itself on the atom:

syntaxstring(p::Proposition; kwargs...) = syntaxstring(atom(p); kwargs...)

Then, the syntaxstring for a given atom can be defined. For example, with String atoms, the function can simply be:

syntaxstring(atom::String; kwargs...) = atom

Warning

The syntaxstring for syntax tokens (e.g., propositions, operators) should not be prefixed/suffixed by whitespaces, as this may cause ambiguities upon parsing. For similar reasons, syntaxstrings should not contain parentheses ('(', ')'), and, when parsing in function notation, commas (','). See also parsebaseformula.

SoleLogics.Proposition — Type

struct Proposition{A} <: AbstractSyntaxToken
    atom::A
end

A proposition, sometimes called a propositional letter (or simply letter), of type Proposition{A} wraps a value atom::A representing a fact which truth can be assessed on a logical interpretation.

Propositions are nullary tokens (i.e, they are at the leaves of a syntax tree). Note that their atom cannot be a Proposition.

See also AbstractSyntaxToken, AbstractInterpretation, check.

Missing docstring.

Missing docstring for negation(p::Proposition). Check Documenter's build log for details.

SoleLogics.AbstractSyntaxStructure — Type

abstract type AbstractSyntaxStructure <: AbstractFormula end

A logical formula unanchored to any logic, and solely represented by its syntactic component. Classically, this structure is implemented as a tree structure (see SyntaxTree); however, the implementation in some cases (e.g., conjuctive/disjuctive normal forms) can differ.

See also tree, SyntaxTree, AbstractFormula, AbstractLogic.

SoleLogics.SyntaxTree — Type

struct SyntaxTree{
    T<:AbstractSyntaxToken,
} <: AbstractSyntaxStructure
    token::T
    children::NTuple{N,SyntaxTree} where {N}
end

A syntax tree encoding a logical formula. Each node of the syntax tree holds a token::T, and has as many children as the arity of the token.

This implementation is arity-compliant, in that, upon construction, the arity is checked against the number of children provided.

See also token, children, tokentype, tokens, operators, propositions, ntokens, npropositions, height, tokenstype, operatorstype, propositionstype, AbstractSyntaxToken, arity, Proposition, Operator.

Base.in — Method

Base.in(tok::AbstractSyntaxToken, f::AbstractFormula)::Bool

Return whether a syntax token appears in a formula.

See also AbstractSyntaxToken.

Missing docstring.

Missing docstring for ntokens. Check Documenter's build log for details.

SoleLogics.tree — Function

tree(f::AbstractFormula)::SyntaxTree

Extract the SyntaxTree representation of a formula (equivalent to Base.convert(SyntaxTree, f)).

See also SyntaxTree, AbstractSyntaxStructure. AbstractFormula,