Introduction

HeytingTruth

HeytingAlgebra

A simple example

Introduction

SoleLogics also provides tools to work with many-valued logics (e.g., fuzzy logics), that is, logics with more truth values other than the classical Boolean ones ⊤ and ⊥. With many-valued logics, the truth values are part of a bounded lattice encoding a partial order between them, encoding a truer than relation.

The main reference, here, is Many-Valued Modal Logics by Melvin Fitting.

HeytingTruth

SoleLogics.HeytingTruth — Type

struct HeytingTruth <: Truth
    label::String
    index::Int
end

A truth value of a Heyting algebra. Heyting truth values are represented by a label, and an index corresponding to its position in the domain vector of the associated algebra. Values ⊤ and ⊥ always exist with index 1 and 2, respectively. New values can be easily constructed via the @heytingtruths macro.

See also @heytingtruths, HeytingAlgebra, Truth

Missing docstring.

Missing docstring for heytingtruths. Check Documenter's build log for details.

HeytingAlgebra

SoleLogics.HeytingAlgebra — Type

struct HeytingAlgebra{D<:AbstractVector{HeytingTruth},
                      G<:Graphs.SimpleGraphs.SimpleDiGraph}
    domain::D
    graph::G
    transitiveclosure::G
end

A Heyting algebra, represented explicitly as a domain of truth values, and a graph over them encoding a partial order with specific constraints (see here). ⊤ and ⊥ are always the first and the second element of each algebra, respectively. A copy of the graph under transitive closure is also stored for optimization purposes.

See also @heytingalgebra, HeytingTruth

Missing docstring.

Missing docstring for heytingalgebra. Check Documenter's build log for details.

A simple example

Let's take as an example the simplest meaningful Heyting Algebra, characterised by the 4 values ⊥, α, β, and ⊤, where α and β are both greater than ⊥ but lesser than ⊤.

This could be represented by the following bounded lattice:

   ⊤
 /   \
α     β
 \   /
   ⊥

To declare a new Heyting algebra in SoleLogics, we can use the heytingalgebra macro, which takes as input a tuple containing the symbols representing the new values of the algebra other than ⊥ and ⊤, and the direct relations between these values, with each relation being a tuple (t1, t2) asserting that t1 < t2. The macro will take care of the declaration of HeytingTruths.

!!!info Please note how both the HeytingTruths and the HeytingAlgebra defined in this way are declared as constants in the global scope.

julia> SoleLogics.@heytingalgebra myalgebra (α, β) (⊥, α) (⊥, β) (α, ⊤) (β, ⊤)
HeytingAlgebra(HeytingTruth[⊤, ⊥, α, β], SimpleDiGraph{Int64}(4, [Int64[], [3, 4], [1], [1]], [[3, 4], Int64[], [2], [2]]))

The classical Boolean connectives are extended to the meet (lower greatest bound), join (greater lowest bound) and Heyting implication operations. These are computed via the collatetruth method, which requires the algebra as an additional argument.

!!!info Note how the truth values have no meaning by themselves, and they must always be associated with an algebra!

julia> collatetruth(∧, (α, β), myalgebra)
HeytingTruth: ⊥

julia> collatetruth(∨, (α, β), myalgebra)
HeytingTruth: ⊤

julia>  collatetruth(→, (α, β), myalgebra)
HeytingTruth: β

```