Getting started

Syntax Basics

Logical formulas are synctactical objects representing statements which level of truth can be assessed. Formulas arise from formal grammars (e.g., context-free grammars), and are most commonly represented as syntax trees. At the leaf nodes of a syntax tree are atoms (simple, atomic statements) or truth values (e.g., ⊤, representing truth), while at the internal nodes are logical connectives that allow for the composition of formulas to represent complex concepts.

In SoleLogics, an Atom is a wrapper for any value (accessible via the value method).

arity

Missing docstring.

Missing docstring for Atom. Check Documenter's build log for details.

Missing docstring.

Missing docstring for TruthValue. Check Documenter's build log for details.

Missing docstring.

Missing docstring for AbstractOperator. Check Documenter's build log for details.

Missing docstring.

Missing docstring for SyntaxTree. Check Documenter's build log for details.

Let's review these concepts with a simple example.

julia> p = Atom("p")
Atom{String}("p")

julia> q = Atom("q")
Atom{String}("q")

julia> value(p)
"p"

# Operators are syntax tokens too
julia> CONJUNCTION
∧

# SyntaxTree's are arity-compliant
julia> st = SyntaxTree(CONJUNCTION, p)
ERROR: AssertionError: Cannot instantiate SyntaxTree{SoleLogics.NamedOperator{:∧}} 
with token ∧ of arity 2 and 1 children.

# In fact, the conjunction operator in a syntax tree must have exactly 2 children 
julia> arity(CONJUNCTION)
2

julia> stree = SyntaxTree(CONJUNCTION, (p,q))
SyntaxTree: p ∧ q

# Get the token of the root node
julia> token(st)
∧

# Get the first subtree, containing only an atom
julia> leftree = children(st)[1]; 
SyntaxTree: p

julia> typeof(leftree)
SyntaxTree{Atom{String}}

# Atoms are necessarily at the leaves; in fact their arity is 0
julia> leftree |> token |> arity
0

Missing docstring.

Missing docstring for AbstractSyntaxStructure. Check Documenter's build log for details.

Missing docstring.

Missing docstring for AbstractSyntaxToken. Check Documenter's build log for details.

SoleLogics.arity — Method

arity(::Type{<:AbstractSyntaxToken})::Integer
arity(tok::AbstractSyntaxToken)::Integer = arity(typeof(tok))

Return the arity of a syntax token. The arity of a syntax token is an integer representing the number of allowed children in a SyntaxTree. Tokens with arity equal to 0, 1 or 2 are called nullary, unary and binary, respectively.

See also AbstractSyntaxToken.

Connectives

TODO

Missing docstring.

Missing docstring for SoleLogics.AbstractOperator. Check Documenter's build log for details.

Note

SoleLogics.jl offers the possibility to implement custom operators. To see an in-depth example, please refer to section Customization.

Missing docstring.

Missing docstring for SoleLogics.NamedOperator. Check Documenter's build log for details.

Missing docstring.

Missing docstring for Base.operator_precedence(op::AbstractOperator). Check Documenter's build log for details.

SoleLogics.isrightassociative — Function

isrightassociative(::Type{AbstractOperator})
isrightassociative(o::AbstractOperator) = isrightassociative(typeof(o))

Return whether an AbstractOperator is right associative or no.

Associativity establishes how operators of the same precedence are grouped in the absence of the parentheses.

Conjunction and disjunction are commutative operators, thus, the left associativity case "(p ∧ q) ∧ r" and the right associativity case "p ∧ (q ∧ r)" are equivalent; by convention we consider the latter form. Implication is right associative, meaning that "p → q → r" is grouped as "p → (q → r)".

By default, an operator is right associative.

Examples

julia> isrightassociative(∧)
true

julia> isrightassociative(→)
true

See also AbstractOperator.

SoleLogics.iscommutative — Function

iscommutative(::Type{AbstractOperator})
iscommutative(o::AbstractOperator) = iscommutative(typeof(o))

Return whether an operator is known to be commutative.

Examples

julia> iscommutative(∧)
true

julia> iscommutative(→)
false

Note that nullary and unary operators are considered commutative.

See also AbstractOperator.

Implementation

When implementing a new type for a commutative operator O with arity higher than 1, please provide a method iscommutative(::Type{O}). This can help model checking operations.

To learn more about operators, refer to Propositional Logic and Modal Logic chapters.

Formulas

TODO: the following definition might be unclear, since "evaluation" and "logic" are not concepts already read by the user. When the Logic structure will be removed from SoleLogics, this section will be expanded with semantics-related concepts.

Missing docstring.

Missing docstring for AbstractFormula. Check Documenter's build log for details.

Note

SoleLogics.jl offers the possibility to implement a custom Formula type. To see an in-depth example, please refer to section Customization.

SoleLogics.joinformulas — Function

joinformulas(
    op::AbstractOperator,
    ::NTuple{N,F}
)::F where {N,F<:AbstractFormula}

Return a new formula of type F by composing N formulas of the same type via an operator op. This function allows one to use operators for flexibly composing formulas (see Implementation).

Examples

julia> f = parseformula("◊(p→q)");

julia> p = Atom("p");

julia> ∧(f, p)  # Easy way to compose a formula
SyntaxTree: ◊(p → q) ∧ p

julia> f ∧ ¬p   # Leverage infix notation ;)
SyntaxTree: ◊(p → q) ∧ ¬p

julia> ∧(f, p, ¬p) # Shortcut for ∧(f, ∧(p, ¬p))
SyntaxTree: ◊(p → q) ∧ p ∧ ¬p

Implementation

Upon joinformulas lies a flexible way of using operators for composing formulas and syntax tokens (e.g., atoms), given by methods like the following:

function (op::AbstractOperator)(
    children::NTuple{N,Union{AbstractSyntaxToken,AbstractFormula}},
) where {N}
    ...
end

These allow composing formulas as in ∧(f, ¬p), and in order to access this composition with any newly defined subtype of AbstractFormula, a new method for joinformulas should be defined, together with promotion from/to other AbstractFormulas should be taken care of (see here and here).

Similarly, for allowing a (possibly newly defined) operator to be applied on a number of syntax tokens/formulas that differs from its arity, for any newly defined operator op, new methods similar to the two above should be defined. For example, although ∧ and ∨ are binary, (i.e., have arity equal to 2), compositions such as ∧(f, f2, f3, ...) and ∨(f, f2, f3, ...) can be done thanks to the following two methods that were defined in SoleLogics:

function ∧(
    c1::Union{AbstractSyntaxToken,AbstractFormula},
    c2::Union{AbstractSyntaxToken,AbstractFormula},
    c3::Union{AbstractSyntaxToken,AbstractFormula},
    cs::Union{AbstractSyntaxToken,AbstractFormula}...
)
    return ∧(c1, ∧(c2, c3, cs...))
end
function ∨(
    c1::Union{AbstractSyntaxToken,AbstractFormula},
    c2::Union{AbstractSyntaxToken,AbstractFormula},
    c3::Union{AbstractSyntaxToken,AbstractFormula},
    cs::Union{AbstractSyntaxToken,AbstractFormula}...
)
    return ∨(c1, ∨(c2, c3, cs...))
end

See also AbstractFormula, AbstractOperator.

Missing docstring.

Missing docstring for tokens(f::AbstractFormula). Check Documenter's build log for details.

Missing docstring.

Missing docstring for SoleLogics.height(t::SyntaxTree). Check Documenter's build log for details.

Missing docstring.

Missing docstring for SoleLogics.tree(f::AbstractFormula). Check Documenter's build log for details.

Parsing & Printing

SoleLogics.jl allows you to:

Extract the string representation of a formula (via syntaxstring);

Parse formulas from strings (via parsetree).

These features are highly customizable, and leverage three properties of syntactical tokens: arity, operator precedence (operator_precedence) and operator associativity (isrightassociative).

SoleLogics.syntaxstring — Function

syntaxstring(φ::AbstractFormula; kwargs...)::String
syntaxstring(tok::AbstractSyntaxToken; kwargs...)::String

Produce the string representation of a formula or syntax token by performing a tree traversal of the syntax tree representation of the formula. Note that this representation may introduce redundant parentheses. kwargs can be used to specify how to display syntax tokens/trees under some specific conditions.

The following kwargs are currently supported:

function_notation = false::Bool: when set to true, it forces the use of function notation for binary operators (see here).

remove_redundant_parentheses = true::Bool: when set to false, it prints a syntaxstring where each syntactical element is wrapped in parentheses.

parenthesize_atoms = !remove_redundant_parentheses::Bool: when set to true, it forces the atoms (which are the leafs of a formula's tree structure) to be wrapped in parentheses.

Examples

julia> syntaxstring(parsebaseformula("p∧q∧r∧s∧t"))
"p ∧ q ∧ r ∧ s ∧ t"

julia> syntaxstring(parsebaseformula("p∧q∧r∧s∧t"), function_notation=true)
"∧(p, ∧(q, ∧(r, ∧(s, t))))"

julia> syntaxstring(parsebaseformula("p∧q∧r∧s∧t"), remove_redundant_parentheses=false)
"(p) ∧ ((q) ∧ ((r) ∧ ((s) ∧ (t))))"

julia> syntaxstring(parsebaseformula("p∧q∧r∧s∧t"), remove_redundant_parentheses=true, parenthesize_atoms=true)
"(p) ∧ (q) ∧ (r) ∧ (s) ∧ (t)"

julia> syntaxstring(parsebaseformula("◊((p∧s)→q)"))
"◊((p ∧ s) → q)"

julia> syntaxstring(parsebaseformula("◊((p∧s)→q)"); function_notation = true)
"◊(→(∧(p, s), q))"

See also parsebaseformula, parsetree, SyntaxTree, AbstractSyntaxToken.

Implementation

In the case of a syntax tree, syntaxstring is a recursive function that calls itself on the syntax children of each node. For a correct functioning, the syntaxstring must be defined (including kwargs...) for every newly defined AbstractSyntaxToken (e.g., operators and Atoms), in a way that it produces a unique string representation, since Base.hash and Base.isequal, at least for SyntaxTrees, rely on it.

In particular, for the case of Atoms, the function calls itself on the wrapped value:

syntaxstring(p::Atom; kwargs...) = syntaxstring(value(p); kwargs...)

Then, the syntaxstring for a given value can be defined. For example, with String (or Number) values, it defaults to:

syntaxstring(value::String; kwargs...) = value

Warning

The syntaxstring for syntax tokens (e.g., atoms, operators) should not be prefixed/suffixed by whitespaces, as this may cause ambiguities upon parsing. For similar reasons, syntaxstrings should not contain parentheses ('(', ')'), and, when parsing in function notation, commas (','). See also parsebaseformula.

SoleLogics.parsetree — Function

parsetree(
    expression::String,
    additional_operators::Union{Nothing,Vector{<:AbstractOperator}} = nothing;
    function_notation::Bool = false,
    atom_parser::Base.Callable = Atom{String},
    additional_whitespaces::Vector{Char} = Char[],
    opening_parenthesis::String = "(",
    closing_parenthesis::String = ")",
    arg_delim::String = ","
)

Return a SyntaxTree which is the result of parsing expression via the Shunting yard algorithm. By default, this function is only able to parse operators in SoleLogics.BASE_PARSABLE_OPERATORS (see arguments section); additional operators may be provided as a second argument.

Arguments

expression::String: expression to be parsed;

additional_operators::Vector{<:AbstractOperator}: additional, non-standard operators needed to correctly parse the expression. When left unset, only the operators in SoleLogics.BASE_PARSABLE_OPERATORS are correctly parsed: S, o, l, e, L, o, g, i, c, s, ., A, b, s, t, r, a, c, t, O, p, e, r, a, t, o, r, [, ⊤, ,, , ⊥, ,, , ¬, ,, , ∧, ,, , ∨, ,, , →, ,, , ◊, ,, , □, ,, , ⟨, G, ⟩, ,, , [, G, ], ,, , ⟨, =, ⟩, ,, , [, =, ], ]; note that, in case of clashing syntaxstring's, the provided additional operators will override these.

Keyword Arguments

function_notation::Bool = false: if set to true, the expression is considered in function notation (e.g, "⨁(arg1, arg2)"); otherwise, it is considered in infix notation (e.g, "arg1 ⨁ arg2");

atom_parser::Base.Callable = Atom{String}: a callable to be used for parsing atoms, once they are recognized in the expression. It must return the atom, or the Atom itself;

additional_whitespaces::Vector{Char} = Char[]: characters to be stripped out from each syntax token. For example, if '@' in additional_whitespaces, "¬@p@" is parsed just as "¬p".

opening_parenthesis::String = "(": the string signaling the opening of an expression block;

closing_parenthesis::String = ")": the string signaling the closing of an expression block;

arg_delim::String = ",": when function_notation = true, the string that delimits the different arguments of a function call.

Warning

For a proper functioning, the syntaxstring of any syntax token cannot be prefixed/suffixed by whitespaces. For example, for any operator ⨁, it should hold that syntaxstring(⨁) == strip(syntaxstring(⨁)). Also, syntaxstrings cannot contain special symbols (opening_parenthesis, closing_parenthesis, and arg_delim) as substrings.

Examples

julia> syntaxstring(parsetree("¬p∧q∧(¬s∧¬z)"))
"¬p ∧ q ∧ ¬s ∧ ¬z"

julia> syntaxstring(parsetree("∧(¬p,∧(q,∧(¬s,¬z)))", function_notation=true))
"¬p ∧ q ∧ ¬s ∧ ¬z"

julia> syntaxstring(parsetree("¬1→0"; atom_parser = (x -> Atom{Float64}(parse(Float64, x)))))
"(¬1.0) → 0.0"

See also SyntaxTree, syntaxstring, [].

Grammar

Missing docstring.

Missing docstring for AbstractAlphabet. Check Documenter's build log for details.

Note

SoleLogics.jl offers the possibility to implement a custom AbstractAlphabet concrete type. To see an in-depth example, please refer to section Customization.

Missing docstring.

Missing docstring for atoms(a::AbstractAlphabet). Check Documenter's build log for details.

Missing docstring.

Missing docstring for Base.in(p::Atom, a::AbstractAlphabet). Check Documenter's build log for details.

Missing docstring.

Missing docstring for Base.length(a::AbstractAlphabet). Check Documenter's build log for details.

Missing docstring.

Missing docstring for Base.iterate(a::AbstractAlphabet). Check Documenter's build log for details.

Missing docstring.

Missing docstring for ExplicitAlphabet. Check Documenter's build log for details.

Missing docstring.

Missing docstring for AlphabetOfAny. Check Documenter's build log for details.

Missing docstring.

Missing docstring for AbstractGrammar. Check Documenter's build log for details.

Missing docstring.

Missing docstring for alphabet(g::AbstractGrammar{A} where {A}). Check Documenter's build log for details.

Missing docstring.

Missing docstring for Base.in(t::SyntaxTree, g::AbstractGrammar). Check Documenter's build log for details.

SoleLogics.formulas — Method

formulas(
    g::AbstractGrammar;
    maxdepth::Integer,
    nformulas::Union{Nothing,Integer} = nothing,
    args...
)::Vector{<:SyntaxTree}

Enumerate the formulas produced by a given grammar with a finite and iterable alphabet.

Implementation

Additional args can be used to model the function's behavior. At least these two arguments should be covered:

a nformulas argument can be used to limit the size of the returned Vector;

a maxdepth argument can be used to limit the syntactic component, represented as a syntax tree,

to a given maximum depth;

See also AbstractGrammar, SyntaxTree.

Missing docstring.

Missing docstring for CompleteFlatGrammar{A<:AbstractAlphabet,O<:AbstractOperator}. Check Documenter's build log for details.

Missing docstring.

Missing docstring for formulas(g::CompleteFlatGrammar{A,O} where {A,O}; maxdepth::Integer, nformulas::Union{Nothing,Integer} = nothing). Check Documenter's build log for details.

Semantics

Customization

Lorem ipsum dolor sit amet, consectetur adipiscing elit. Nunc justo justo, finibus ac odio in, tempor fermentum augue. Vivamus ullamcorper lacus eget enim imperdiet, ac lobortis turpis elementum. Fusce non auctor eros. Duis scelerisque auctor volutpat. Morbi non luctus est. Orci varius natoque penatibus et magnis dis parturient montes, nascetur ridiculus mus. Pellentesque porttitor a est sit amet ornare. Aliquam faucibus fringilla imperdiet.