More on Formulas

Linear Forms

Random sampling and generation

Parsing

Utilities

More on Formulas

In this chapter, you are going to learn more about Formula representations that are alternative to syntax trees. As you will see, for example, formulas with specific structure (e.g., normal forms) can be represented in ways that make them more easy to handle, and can lead to great benefits in terms of both computational and memory load.

We proceed by presenting the random formulae generation engine, parsing and some utility function.

Recalling the type hierarchy presented in man-core, it is here enriched with the following new types and structures.

Formula

AnchoredFormula (new)

AbstractSyntaxStructure

LeftmostLinearForm{C<:Connective,SS<:AbstractSyntaxStructure} (new)

Linear Forms

SoleLogics.LeftmostLinearForm — Type

struct LeftmostLinearForm{C<:Connective,SS<:AbstractSyntaxStructure} <: AbstractSyntaxStructure
    children::Vector{<:SS}
end

A syntax structure representing the foldl of a set of other syntax structure of type SS by means of a connective C. This structure enables a structured instantiation of formulas in conjuctive/disjunctive forms, and conjuctive normal form (CNF) or disjunctive normal form (DNF), defined as:

const LeftmostConjunctiveForm{SS<:AbstractSyntaxStructure} = LeftmostLinearForm{typeof(∧),SS}
const LeftmostDisjunctiveForm{SS<:AbstractSyntaxStructure} = LeftmostLinearForm{typeof(∨),SS}

const CNF{SS<:AbstractSyntaxStructure} = LeftmostLinearForm{typeof(∧),LeftmostLinearForm{typeof(∨),SS}}
const DNF{SS<:AbstractSyntaxStructure} = LeftmostLinearForm{typeof(∨),LeftmostLinearForm{typeof(∧),SS}}

Examples

julia> LeftmostLinearForm(→, parseformula.(["p", "q", "r"]))
LeftmostLinearForm{SoleLogics.NamedConnective{:→},Atom{String}}
    "(p) → (q) → (r)"

julia> LeftmostConjunctiveForm(parseformula.(["¬p", "q", "¬r"]))
LeftmostLinearForm{SoleLogics.NamedConnective{:∧},SyntaxTree}
    "(¬p) ∧ (q) ∧ (¬r)"

julia> LeftmostDisjunctiveForm{Literal}([Literal(false, Atom("p")), Literal(true, Atom("q")), Literal(false, Atom("r"))])
LeftmostLinearForm{SoleLogics.NamedConnective{:∨},Literal}
    "(¬p) ∨ (q) ∨ (¬r)"

julia> LeftmostDisjunctiveForm([LeftmostConjunctiveForm(parseformula.(["¬p", "q", "¬r"]))]) isa SoleLogics.DNF
true

See also AbstractSyntaxStructure, SyntaxTree, LeftmostConjunctiveForm, LeftmostDisjunctiveForm, Literal.

SoleLogics.LeftmostConjunctiveForm — Type

LeftmostConjunctiveForm{SS<:AbstractSyntaxStructure} = LeftmostLinearForm{typeof(∧),SS}

Specific instantiation of a LeftmostLinearForm, where Connectives are all CONJUNCTIONs.

See also AbstractSyntaxStructure, Connective, LeftmostLinearForm, CONJUNCTION.

SoleLogics.LeftmostDisjunctiveForm — Type

LeftmostDisjunctiveForm{SS<:AbstractSyntaxStructure} = LeftmostLinearForm{typeof(∨),SS}

Specific instantiation of a LeftmostLinearForm, where Connectives are all DISJUNCTIONs.

See also AbstractSyntaxStructure, Connective, LeftmostLinearForm, DISJUNCTION.

SoleLogics.CNF — Type

CNF{SS<:AbstractSyntaxStructure} = LeftmostConjunctiveForm{LeftmostDisjunctiveForm{SS}}

Conjunctive Normal Form of an AbstractSyntaxStructure.

See also AbstractSyntaxStructure, LeftmostConjunctiveForm, LeftmostDisjunctiveForm, CONJUNCTION, DISJUNCTION.

SoleLogics.DNF — Type

DNF{SS<:AbstractSyntaxStructure} = LeftmostConjunctiveForm{LeftmostConjunctiveForm{SS}}

Disjunctive Normal Form of an AbstractSyntaxStructure.

See also AbstractSyntaxStructure, LeftmostConjunctiveForm, LeftmostDisjunctiveForm, CONJUNCTION, DISJUNCTION.

SoleLogics.AnchoredFormula — Type

struct AnchoredFormula{L<:AbstractLogic} <: Formula
    _logic::Base.RefValue{L}
    synstruct::AbstractSyntaxStructure
end

A formula anchored to a logic of type L, and wrapping a syntax structure. The structure encodes a formula belonging to the grammar of the logic, and the truth of the formula can be evaluated on interpretations of the same logic. Note that, here, the logic is represented by a reference.

Upon construction, the logic can be passed either directly, or via a RefValue. Additionally, the following keyword arguments may be specified:

check_atoms::Bool = false: whether to perform or not a check that the atoms belong to the alphabet of the logic;

check_tree::Bool = false: whether to perform or not a check that the formula's syntactic structure honors the grammar (includes the check performed with check_atoms = true);

Cool feature: a AnchoredFormula can be used for instating other formulas of the same logic. See the examples.

Examples

julia> φ = parsebaseformula("◊(p→q)");

julia> f2 = φ(parseformula("p"));

julia> syntaxstring(φ)
"◊(→(p, q))"

julia> syntaxstring(f2)
"p"

julia> @assert logic(φ) == logic(f2)

julia> @assert ◊ in operators(logic(f2))

julia> @assert ◊ isa operatorstype(logic(f2))

See also AbstractLogic, logic, SyntaxToken, SyntaxBranch, tree.

SoleLogics.logic — Method

logic(φ::AnchoredFormula)::AbstractLogic

Return the logic of an anchored formula

See AnchoredFormula.

SoleLogics.synstruct — Method

synstruct(φ::AnchoredFormula)::AbstractSyntaxStructure

Return the syntactic component of an anchored formula.

See AnchoredFormula.

SoleLogics.baseformula — Method

function baseformula(
    φ::Formula;
    infer_logic = true,
    additional_operators::Union{Nothing,Vector{<:Operator}} = nothing,
    kwargs...
)

Attempt at instantiating a AnchoredFormula from a syntax token/formula, by inferring the logic it belongs to. If infer_logic is true, then a canonical logic (e.g., propositional logic with all the BASE_PROPOSITIONAL_CONNECTIVES) is inferred; if it's false, then a logic with exactly the operators appearing in the syntax tree, plus the additional_operators is instantiated.

Examples

julia> t = parseformula("◊((p∧q)→r)");

julia> unique(operators(logic(SoleLogics.baseformula(t))))
3-element Vector{Union{SoleLogics.NamedConnective{:→}, SoleLogics.NamedConnective{:◊}, SoleLogics.NamedConnective{:∧}}}:
 ∧
 ◊
 →

julia> unique(operators(logic(SoleLogics.baseformula(t; additional_operators = SoleLogics.BASE_MODAL_CONNECTIVES))))
8-element Vector{Union{SoleLogics.BottomOperator, SoleLogics.NamedConnective{:¬}, SoleLogics.NamedConnective{:∧}, SoleLogics.NamedConnective{:∨}, SoleLogics.NamedConnective{:→}, SoleLogics.NamedConnective{:◊}, SoleLogics.NamedConnective{:□}, SoleLogics.TopOperator}}:
 ¬
 ∧
 ∨
 →
 ◊
 □

SoleLogics.parsebaseformula — Function

parsebaseformula(
    expr::String,
    additional_operators::Union{Nothing,Vector{<:Operator}} = nothing;
    operators::Union{Nothing,Vector{<:Operator}},
    grammar::Union{Nothing,AbstractGrammar} = nothing,
    algebra::Union{Nothing,AbstractAlgebra} = nothing,
    kwargs...
)::AnchoredFormula

Return a AnchoredFormula which is the result of parsing an expression via the Shunting yard algorithm. By default, this function is only able to parse operators in SoleLogics.BASE_PARSABLE_CONNECTIVES; additional operators may be provided as a second argument.

The grammar and algebra of the associated logic is inferred using the baseformula function from the operators encountered in the expression, and those in additional_operators.

See parseformula, baseformula, BASE_PARSABLE_CONNECTIVES.

Random sampling and generation

Base.rand — Method

Base.rand(
    [rng::AbstractRNG = Random.GLOBAL_RNG,]
    alphabet::AbstractAlphabet,
    args...;
    kwargs...
)::Atom

Base.rand(
    [rng::AbstractRNG = Random.GLOBAL_RNG,]
    height::Integer,
    g::CompleteFlatGrammar,
    args...
)

Base.rand(
    height::Integer,
    connectives::Union{AbstractVector{<:Operator},AbstractVector{<:Connective}},
    atoms::Union{AbstractVector{<:Atom},AbstractAlphabet},
    truthvalues::Union{Nothing,AbstractVector{<:Truth},AbstractAlgebra} = nothing,
    args...;
    rng::Union{Integer,AbstractRNG} = Random.GLOBAL_RNG,
    kwargs...
)

Randomly sample an atom from an alphabet, according to a uniform distribution.

Implementation

If the alphabet is finite, the function defaults to rand(rng, atoms(alphabet)); otherwise, it must be implemented, and additional keyword arguments should be provided in order to limit the (otherwise infinite) sampling domain.

See also AbstractAlphabet.

Missing docstring.

Missing docstring for StatsBase.sample(alphabet::AbstractAlphabet, weights::AbstractWeights, args...; kwargs...). Check Documenter's build log for details.

Missing docstring.

Missing docstring for randformula(height::Integer, alphabet, operators::AbstractVector; rng::Union{Integer,AbstractRNG} = Random.GLOBAL_RNG). Check Documenter's build log for details.

Parsing

SoleLogics.BASE_PARSABLE_CONNECTIVES — Constant

const BASE_PARSABLE_CONNECTIVES = SoleLogics.Syntactical[¬, ∧, ∨, →, ◊, □, ⟨G⟩, [G], ⟨=⟩, [=], ⊤, ⊥]

Vector of (standard) operators that are automatically taken care of when parsing. These are ¬, ∧, ∨, →, ◊, □, ⟨G⟩, [G], ⟨=⟩, [=], ⊤ and ⊥.

See also parseformula.

SoleLogics.parseformula — Function

parseformula(expr::String, additional_operators = nothing; kwargs...)

parseformula(
    F::Type{<:SyntaxTree},
    expr::String,
    additional_operators::Union{Nothing,AbstractVector} = nothing;
    function_notation::Bool = false,
    atom_parser::Base.Callable = Atom{String},
    additional_whitespaces::Vector{Char} = Char[],
    opening_parenthesis::String = DEFAULT_OPENING_PARENTHESIS,
    closing_parenthesis::String = DEFAULT_CLOSING_PARENTHESIS,
    arg_delim::String = DEFAULT_ARG_DELIM
)::F

parseformula(F::Type{<:Formula}, expr::String, additional_operators = nothing; kwargs...)
parseformula(F::Type{<:SyntaxTree}, expr::String, logic::AbstractLogic; kwargs...)

Parse a formula of type F from a string expression (its syntaxstring). When F is not specified, it defaults to SyntaxTree.

By default, this function is only able to parse operators in SoleLogics.BASE_PARSABLE_CONNECTIVES (e.g., ¬, ∧, ∨ and →); additional, non-standard operators may be provided as a vector additional_operators, and their syntaxstrings will be used for parsing them. Note that, in case of clashing syntaxstrings, the provided additional operators will override the standard ones.

When parsing SyntaxTrees, the Shunting yard algorithm is used, and the method allows the following keywords arguments.

Keyword Arguments

function_notation::Bool = false: if set to true, the expression is considered in function notation (e.g., "⨁(arg1, arg2)"); otherwise, it is considered in infix notation (e.g., "arg1 ⨁ arg2");

atom_parser::Base.Callable = Atom{String}: a callable to be used for parsing atoms, once they are recognized in the expression. It must return the atom, or the Atom itself;

additional_whitespaces::Vector{Char} = Char[]: characters to be stripped out from each syntax token. For example, if '@' in additional_whitespaces, "¬@p@" is parsed just as "¬p".

opening_parenthesis::String = "(": the string signaling the opening of an expression block;

closing_parenthesis::String = ")": the string signaling the closing of an expression block;

arg_delim::String = ",": when function_notation = true, the string that delimits the different arguments of a function call.

Warning

For a proper functioning, the syntaxstring of any syntax token cannot be prefixed/suffixed by whitespaces. For example, for any operator ⨁, it should hold that syntaxstring(⨁) == strip(syntaxstring(⨁)). Also, syntaxstrings cannot contain special symbols (opening_parenthesis, closing_parenthesis, and arg_delim) as substrings.

Examples

julia> syntaxstring(parseformula("¬p∧q∧(¬s∧¬z)"))
"¬p ∧ q ∧ ¬s ∧ ¬z"

julia> syntaxstring(parseformula("∧(¬p,∧(q,∧(¬s,¬z)))", function_notation=true))
"¬p ∧ q ∧ ¬s ∧ ¬z"

julia> syntaxstring(parseformula("¬1→0"; atom_parser = (x -> Atom{Float64}(parse(Float64, x)))))
"(¬1.0) → 0.0"

Note

For any Formula type F, this function should be the inverse of syntaxstring; that is, if φ::F then the following should hold, for at least some args, and for every kwargs allowing correct parsing: φ == parseformula(F, syntaxstring(φ, args...; kwargs...), args...; kwargs...).

See also SyntaxTree, BASE_PARSABLE_CONNECTIVES, syntaxstring.

Utilities

SoleLogics.treewalk — Method

treewalk(
    st::SyntaxTree,
    args...;
    rng::AbstractRNG = Random.GLOBAL_RNG,
    criterion::Function = ntokens,
    toleaf::Bool = true,
    returnnode::Bool = false,
    transformnode::Function = nothing
)::SyntaxTree

Return a subtree of syntax tree, by following these options:

criterion: function used to compute the probability of stopping at a random node;

returnnode: true if only the subtree is to be returned;

transformnode: function that will be applied to the chosen subtree.

SoleLogics.subformulas — Method

subformulas(f::Formula; sorted=true)

Return all sub-formulas (sorted by size when sorted=true) of a given formula.

Examples

julia> syntaxstring.(SoleLogics.subformulas(parseformula("◊((p∧q)→r)")))
6-element Vector{String}:
 "p"
 "q"
 "r"
 "p ∧ q"
 "◊(p ∧ q)"
 "(◊(p ∧ q)) → r"

See also SyntaxTree), Formula.

SoleLogics.normalize — Method

normalize(
    f::Formula;
    remove_boxes = true,
    reduce_negations = true,
    allow_atom_flipping = true
)

Return a modified version of a given formula, that has the same semantics but different syntax. This is useful when dealing with the truth of many (possibly similar) formulas; for example, when performing model checking. BEWARE: it currently assumes the underlying algebra is Boolean!

Arguments

f::Formula: when set to true, the formula;

remove_boxes::Bool: remove all (non-relational and relational) box operators by using the equivalence ◊φ ≡ ¬□¬φ. Note: this assumes an underlying Boolean algebra.

reduce_negations::Bool: when set to true, attempts at reducing the number of negations by appling some transformation rules (e.g., De Morgan's laws). Note: this assumes an underlying Boolean algebra.

allow_atom_flipping::Bool: when set to true, together with reduce_negations=true, this may cause the negation of an atom to be replaced with the its dual atom.

Examples

julia> f = parseformula("□¬((p∧¬q)→r)∧⊤");

julia> syntaxstring(f)
"□¬((p ∧ ¬q) → r) ∧ ⊤"

julia> syntaxstring(SoleLogics.normalize(f; profile = :modelchecking, allow_atom_flipping = false))
"¬◊(q ∨ ¬p ∨ r)"

julia> syntaxstring(SoleLogics.normalize(f; profile = :readability, allow_atom_flipping = false))
"□(¬r ∧ p ∧ ¬q)"

See also SyntaxTree), Formula.

SoleLogics.isgrounded — Method

isgrounded(f::Formula)::Bool

Return true if the formula is grounded, that is, if it can be inferred from its syntactic structure that, given any frame-based model, the truth value of the formula is the same on every world.

Examples

julia> f = parseformula("⟨G⟩p → [G]q");

julia> syntaxstring(f)
"(⟨G⟩p) → ([G]q)"

julia> SoleLogics.isgrounded(f)
true

See also isgrounding), SyntaxTree), Formula.