Getting started

Syntax Basics

Logical formulas are syntactical objects representing statements which level of truth can be assessed. Formulas arise from formal grammars (e.g., context-free grammars), and are most commonly represented as syntax trees. At the leaf nodes of a syntax tree are atoms (simple, atomic statements) or truth values (e.g., ⊤, representing truth), while at the internal nodes are logical connectives that allow for the composition of formulas to represent complex concepts.

In SoleLogics, an Atom is a wrapper for any value (accessible via the value method). TODO arity

Missing docstring.

Missing docstring for Atom. Check Documenter's build log for details.

Missing docstring.

Missing docstring for Truth. Check Documenter's build log for details.

Missing docstring.

Missing docstring for Connective. Check Documenter's build log for details.

Missing docstring.

Missing docstring for SyntaxBranch. Check Documenter's build log for details.

Let's review these concepts with a simple example.

julia> p = Atom("p")
Atom{String}("p")

julia> q = Atom("q")
Atom{String}("q")

julia> value(p)
"p"

# Operators are syntax tokens too
julia> CONJUNCTION
∧

# SyntaxBranch's are arity-compliant
julia> st = SyntaxBranch(CONJUNCTION, p)
ERROR: AssertionError: Cannot instantiate SyntaxBranch{SoleLogics.NamedConnective{:∧}} 
with token ∧ of arity 2 and 1 children.

# In fact, the conjunction operator in a syntax tree must have exactly 2 children 
julia> arity(CONJUNCTION)
2

julia> stree = SyntaxBranch(CONJUNCTION, (p,q))
SyntaxBranch: p ∧ q

# Get the token of the root node
julia> token(st)
∧

# Get the first subtree, containing only an atom
julia> leftree = children(st)[1]; 
SyntaxBranch: p

julia> typeof(leftree)
SyntaxBranch{Atom{String}}

# Atoms are necessarily at the leaves; in fact their arity is 0
julia> leftree |> token |> arity
0

Missing docstring.

Missing docstring for AbstractSyntaxStructure. Check Documenter's build log for details.

Missing docstring.

Missing docstring for SyntaxToken. Check Documenter's build log for details.

Missing docstring.

Missing docstring for SoleLogics.arity(::SyntaxToken). Check Documenter's build log for details.

Connectives

TODO

Missing docstring.

Missing docstring for SoleLogics.Connective. Check Documenter's build log for details.

Note

SoleLogics.jl offers the possibility to implement custom operators. To see an in-depth example, please refer to section Customization.

Missing docstring.

Missing docstring for SoleLogics.NamedConnective. Check Documenter's build log for details.

SoleLogics.precedence — Function

precedence(c::Connective)

Return the precedence of a binary connective.

When using infix notation, and in the absence of parentheses, precedence establishes how binary connectives are interpreted. A precedence value is a standard integer, and connectives with high precedence take precedence over connectives with lower precedences. This affects how formulas are shown (via syntaxstring) and parsed (via parseformula).

By default, the value for a NamedConnective is derived from the Base.operator_precedence of its symbol (name); there are some exceptions (e.g., ¬). Because of this, when dealing with a custom connective ⊙, it will be the case that parseformula("p ⊙ q ∧ r") == (@synexpr p ⊙ q ∧ r).

It is possible to assign a specific precedence to a connective type C by providing a method Base.operator_precedence(::C).

Examples

julia> precedence(∧) == Base.operator_precedence(:∧)
true

julia> precedence(∧), precedence(∨), precedence(→)
∨(12, 11, 4)

julia> syntaxstring(parseformula("¬a ∧ b ∧ c"))
"¬a ∧ b ∧ c"

julia> syntaxstring(parseformula("¬a → b ∧ c"))
"(¬a) → (b ∧ c)"

julia> syntaxstring(parseformula("a ∧ b → c ∧ d"))
"(a ∧ b) → (c ∧ d)"

See also associativity, Connective.

SoleLogics.associativity — Function

associativity(::Connective)

Return whether a (binary) connective is right-associative.

When using infix notation, and in the absence of parentheses, associativity establishes how binary connectives of the sameprecedenceare interpreted. This affects how formulas are shown (viasyntaxstring) and parsed (viaparseformula`).

By default, the value for a NamedConnective is derived from the Base.operator_precedence of its symbol (name); thus, for example, most connectives are left-associative (e.g., ∧ and ∨), while → is right-associative. Because of this, when dealing with a custom connective ⊙, it will be the case that parseformula("p ⊙ q ∧ r") == (@synexpr p ⊙ q ∧ r).

Examples

julia> associativity(∧)
:left

julia> associativity(→)
:right

julia> syntaxstring(parseformula("p → q → r"); remove_redundant_parentheses = false)
"p → (q → r)"

julia> syntaxstring(parseformula("p ∧ q ∨ r"); remove_redundant_parentheses = false)
"(p ∧ q) ∨ r"

See also Connective, parseformula, precedence, syntaxstring.

SoleLogics.iscommutative — Function

iscommutative(c::Connective)

Return whether a connective is known to be commutative.

Examples

julia> iscommutative(∧)
true

julia> iscommutative(→)
false

Note that nullary and unary connectives are considered commutative.

See also Connective.

Implementation

When implementing a new type for a commutative connective C with arity higher than 1, please provide a method iscommutative(::C). This can help model checking operations.

See also Connective.

To learn more about operators, refer to Propositional Logic and Modal Logic chapters.

Formulas

TODO: the following definition might be unclear, since "evaluation" and "logic" are not concepts already read by the user. When the Logic structure will be removed from SoleLogics, this section will be expanded with semantics-related concepts.

Missing docstring.

Missing docstring for Formula. Check Documenter's build log for details.

Note

SoleLogics.jl offers the possibility to implement a custom Formula subtype. To see an in-depth example, please refer to section Customization.

SoleLogics.composeformulas — Function

composeformulas(c::Connective, φs::NTuple{N,F})::F where {N,F<:Formula}

Return a new formula of type F by composing N formulas of the same type via a connective c. This function allows one to use connectives for flexibly composing formulas (see Implementation section).

Examples

julia> f = parseformula("◊(p→q)");

julia> p = Atom("p");

julia> ∧(f, p)  # Easy way to compose a formula
SyntaxBranch: ◊(p → q) ∧ p

julia> f ∧ ¬p   # Leverage infix notation ;) (see https://stackoverflow.com/a/60321302/5646732)
SyntaxBranch: ◊(p → q) ∧ ¬p

julia> ∧(f, p, ¬p) # Shortcut for ∧(f, ∧(p, ¬p))
SyntaxBranch: ◊(p → q) ∧ p ∧ ¬p

Implementation

Upon composeformulas lies a flexible way of using connectives for composing formulas and syntax tokens (e.g., atoms), given by methods like the following:

function (c::Connective)(φs::NTuple{N,Formula}) where {N}
    ...
end

These allow composing formulas as in ∧(f, ¬p), and in order to access this composition with any newly defined subtype of Formula, a new method for composeformulas should be defined, together with promotion from/to other Formulas should be taken care of (see here and here).

Similarly, for allowing a (possibly newly defined) connective to be applied on a number of syntax tokens/formulas that differs from its arity, for any newly defined connective c, new methods similar to the two above should be defined. For example, although ∧ and ∨ are binary, (i.e., have arity equal to 2), compositions such as ∧(f, f2, f3, ...) and ∨(f, f2, f3, ...) can be done thanks to the following two methods that were defined in SoleLogics:

function ∧(
    c1::Formula,
    c2::Formula,
    c3::Formula,
    cs::Formula...
)
    return ∧(c1, ∧(c2, c3, cs...))
end
function ∨(
    c1::Formula,
    c2::Formula,
    c3::Formula,
    cs::Formula...
)
    return ∨(c1, ∨(c2, c3, cs...))
end

Note

To allow for the composition of Formulas of different types, promotion rules should be provided.

See also Formula, Connective.

Missing docstring.

Missing docstring for tokens(::Formula). Check Documenter's build log for details.

SoleLogics.height — Method

height(φ::Formula)::Integer

Return the height of a formula in its syntax tree representation.

See also SyntaxTree.

Missing docstring.

Missing docstring for SoleLogics.tree(::Formula). Check Documenter's build log for details.

Parsing & Printing

SoleLogics.jl allows you to:

Extract the string representation of a formula (via syntaxstring);

Parse formulas from strings (via parseformula).

These features are highly customizable, and leverage three properties of syntactical tokens: arity, connective precedence and connective associativity.

SoleLogics.syntaxstring — Function

syntaxstring(s::Syntactical; kwargs...)::String

Return the string representation of any syntactic object (e.g., Formula, SyntaxTree, SyntaxToken, Atom, Truth, etc). Note that this representation may introduce redundant parentheses. kwargs can be used to specify how to display syntax tokens/trees under some specific conditions.

The following kwargs are currently supported:

function_notation = false::Bool: when set to true, it forces the use of function notation for binary operators (see here).

remove_redundant_parentheses = true::Bool: when set to false, it prints a syntaxstring where each syntactical element is wrapped in parentheses.

parenthesize_atoms = !remove_redundant_parentheses::Bool: when set to true, it forces the atoms (which are the leaves of a formula's tree structure) to be wrapped in parentheses.

Examples

julia> syntaxstring(parseformula("p∧q∧r∧s∧t"))
"p ∧ q ∧ r ∧ s ∧ t"

julia> syntaxstring(parseformula("p∧q∧r∧s∧t"), function_notation=true)
"∧(∧(∧(∧(p, q), r), s), t)"

julia> syntaxstring(parseformula("p∧q∧r∧s∧t"), remove_redundant_parentheses=false)
"((((p) ∧ (q)) ∧ (r)) ∧ (s)) ∧ (t)"

julia> syntaxstring(parseformula("p∧q∧r∧s∧t"), remove_redundant_parentheses=true, parenthesize_atoms=true)
"(p) ∧ (q) ∧ (r) ∧ (s) ∧ (t)"

julia> syntaxstring(parseformula("◊((p∧s)→q)"))
"◊((p ∧ s) → q)"

julia> syntaxstring(parseformula("◊((p∧s)→q)"); function_notation = true)
"◊(→(∧(p, s), q))"

See also parseformula, SyntaxBranch, SyntaxToken.

Implementation

In the case of a syntax tree, syntaxstring is a recursive function that calls itself on the syntax children of each node. For a correct functioning, the syntaxstring must be defined (including kwargs...) for every newly defined SyntaxToken (e.g., SyntaxLeafs, that is, Atoms and Truth values, and Operators), in a way that it produces a unique string representation, since Base.hash and Base.isequal, at least for SyntaxBranchs, rely on it.

In particular, for the case of Atoms, the function calls itself on the wrapped value:

syntaxstring(a::Atom; kwargs...) = syntaxstring(value(a); kwargs...)

The syntaxstring for any value defaults to its string representation, but it can be defined by defining the appropriate syntaxstring method.

Warning

The syntaxstring for syntax tokens (e.g., atoms, operators) should not be prefixed/suffixed by whitespaces, as this may cause ambiguities upon parsing. For similar reasons, syntaxstrings should not contain parentheses ('(', ')'), and, when parsing in function notation, commas (',').

See also SyntaxLeaf, Operator, parseformula.

SoleLogics.parseformula — Function

parseformula(expr::String, additional_operators = nothing; kwargs...)
parseformula(F::Type{<:Formula}, expr::String, additional_operators = nothing; kwargs...)

Parse a formula of type F from a string expression (its syntaxstring). When F is not specified, it defaults to SyntaxTree.

By default, this function is only able to parse operators in SoleLogics.BASE_PARSABLE_OPERATORS (e.g., ¬, ∧, ∨ and →); additional, non-standard operators may be provided as a vector additional_operators, and their syntaxstring's will be used for parsing them. Note that, in case of clashing syntaxstring's, the provided additional operators will override the standard ones.

When parsing SyntaxTrees, the Shunting yard algorithm is used, and the method allows the following keywords arguments.

Keyword Arguments

function_notation::Bool = false: if set to true, the expression is considered in function notation (e.g., "⨁(arg1, arg2)"); otherwise, it is considered in infix notation (e.g., "arg1 ⨁ arg2");

atom_parser::Base.Callable = Atom{String}: a callable to be used for parsing atoms, once they are recognized in the expression. It must return the atom, or the Atom itself;

additional_whitespaces::Vector{Char} = Char[]: characters to be stripped out from each syntax token. For example, if '@' in additional_whitespaces, "¬@p@" is parsed just as "¬p".

opening_parenthesis::String = "(": the string signaling the opening of an expression block;

closing_parenthesis::String = ")": the string signaling the closing of an expression block;

arg_delim::String = ",": when function_notation = true, the string that delimits the different arguments of a function call.

Warning

For a proper functioning, the syntaxstring of any syntax token cannot be prefixed/suffixed by whitespaces. For example, for any operator ⨁, it should hold that syntaxstring(⨁) == strip(syntaxstring(⨁)). Also, syntaxstrings cannot contain special symbols (opening_parenthesis, closing_parenthesis, and arg_delim) as substrings.

Examples

julia> syntaxstring(parseformula("¬p∧q∧(¬s∧¬z)"))
"¬p ∧ q ∧ ¬s ∧ ¬z"

julia> syntaxstring(parseformula("∧(¬p,∧(q,∧(¬s,¬z)))", function_notation=true))
"¬p ∧ q ∧ ¬s ∧ ¬z"

julia> syntaxstring(parseformula("¬1→0"; atom_parser = (x -> Atom{Float64}(parse(Float64, x)))))
"(¬1.0) → 0.0"

Note

For any Formula type F, this function should be the inverse of syntaxstring; that is, if φ::F then the following should hold, for at least some args, and for every kwargs allowing correct parsing: φ == parseformula(F, syntaxstring(φ, args...; kwargs...), args...; kwargs...).

See also SyntaxTree, BASE_PARSABLE_OPERATORS, syntaxstring.

Grammar

Missing docstring.

Missing docstring for AbstractAlphabet. Check Documenter's build log for details.

Note

SoleLogics.jl offers the possibility to implement a custom AbstractAlphabet concrete type. To see an in-depth example, please refer to section Customization.

Missing docstring.

Missing docstring for atoms(a::AbstractAlphabet). Check Documenter's build log for details.

Missing docstring.

Missing docstring for Base.in(p::Atom, a::AbstractAlphabet). Check Documenter's build log for details.

Missing docstring.

Missing docstring for Base.length(a::AbstractAlphabet). Check Documenter's build log for details.

Missing docstring.

Missing docstring for Base.iterate(a::AbstractAlphabet). Check Documenter's build log for details.

Missing docstring.

Missing docstring for ExplicitAlphabet. Check Documenter's build log for details.

Missing docstring.

Missing docstring for AlphabetOfAny. Check Documenter's build log for details.

Missing docstring.

Missing docstring for AbstractGrammar. Check Documenter's build log for details.

SoleLogics.alphabet — Method

alphabet(g::AbstractGrammar{V} where {V})::V

Return the propositional alphabet of a grammar.

See also AbstractAlphabet, AbstractGrammar.

Base.in — Method

Base.in(φ::SyntaxTree, g::AbstractGrammar)::Bool

Return whether a SyntaxTree, belongs to a grammar.

See also AbstractGrammar, SyntaxTree.

SoleLogics.formulas — Method

formulas(
    g::AbstractGrammar;
    maxdepth::Integer,
    nformulas::Union{Nothing,Integer} = nothing,
    args...
)::Vector{<:SyntaxBranch}

Enumerate the formulas produced by a given grammar with a finite and iterable alphabet.

Implementation

Additional args can be used to model the function's behavior. At least these two arguments should be covered:

a nformulas argument can be used to limit the size of the returned Vector;

a maxdepth argument can be used to limit the syntactic component, represented as a syntax tree,

to a given maximum depth;

See also AbstractGrammar, SyntaxBranch.

Missing docstring.

Missing docstring for CompleteFlatGrammar{A<:AbstractAlphabet,O<:Operator}. Check Documenter's build log for details.

SoleLogics.formulas — Method

formulas(
    g::CompleteFlatGrammar{V,O} where {V,O};
    maxdepth::Integer,
    nformulas::Union{Nothing,Integer} = nothing
)::Vector{SyntaxBranch}

Generate all formulas whose SyntaxBranchs that are not taller than a given maxdepth.

See also AbstractGrammar, SyntaxBranch.

Semantics

Customization

Lorem ipsum dolor sit amet, consectetur adipiscing elit. Nunc justo justo, finibus ac odio in, tempor fermentum augue. Vivamus ullamcorper lacus eget enim imperdiet, ac lobortis turpis elementum. Fusce non auctor eros. Duis scelerisque auctor volutpat. Morbi non luctus est. Orci varius natoque penatibus et magnis dis parturient montes, nascetur ridiculus mus. Pellentesque porttitor a est sit amet ornare. Aliquam faucibus fringilla imperdiet.